Factorize tests into a single file
